requirements.txt
# ====================================================================
# REQUIREMENTS - PLATAFORMA PREUNIVERSITARIA
# ====================================================================
# 
# INSTALACIÓN COMPLETA:
# pip install -r requirements_nuevo.txt
#
# INSTALACIÓN POR CATEGORÍAS:
# pip install flask jinja2 werkzeug          # Solo web
# pip install pathlib2 regex python-dateutil # Solo core
# pip install pytest black flake8            # Solo desarrollo
#
# ====================================================================

# ============================================
# DEPENDENCIAS WEB (FLASK APPLICATION)
# ============================================
# Instalar: pip install flask
Flask==2.3.3                    # Framework web principal
Werkzeug==2.3.7                 # Servidor WSGI para Flask
Jinja2==3.1.2                   # Motor de plantillas HTML
MarkupSafe==2.1.3               # Escape seguro para templates
itsdangerous==2.1.2             # Firma criptográfica para sessions
click==8.1.7                    # Interface de línea de comandos
blinker==1.6.3                  # Sistema de señales para Flask

# ============================================
# SISTEMA DE AUTENTICACIÓN
# ============================================
# Instalar: pip install flask-login flask-sqlalchemy werkzeug
Flask-Login==0.6.3              # Sistema de autenticación para Flask
Flask-SQLAlchemy==3.0.5         # ORM para bases de datos
Werkzeug==2.3.7                 # Utilidades de seguridad (ya incluido arriba)

# ============================================
# AUTENTICACIÓN CON GOOGLE OAUTH
# ============================================
# Instalar: pip install requests-oauthlib
requests-oauthlib>=1.3.1        # Cliente OAuth para Google
google-auth>=2.17.3             # Autenticación de Google
google-auth-oauthlib>=1.0.0     # OAuth para Google
google-auth-httplib2>=0.1.0     # HTTP client para Google APIs

# ============================================
# DEPENDENCIAS CORE (PROCESAMIENTO)
# ============================================
# Instalar: pip install pathlib2 regex
pathlib2>=2.3.7                 # Manejo avanzado de rutas
typing-extensions>=4.0.0        # Extensiones de tipado
regex>=2022.1.18                # Expresiones regulares avanzadas
python-dateutil>=2.8.2          # Manejo inteligente de fechas

# ============================================
# PROCESAMIENTO DE DATOS
# ============================================
# Instalar: pip install jsonschema ijson
jsonschema>=4.17.0              # Validación de esquemas JSON
ijson>=3.1.4                    # Procesamiento eficiente de JSON grandes
unidecode>=1.3.6                # Conversión Unicode a ASCII

# ============================================
# LOGGING Y CONFIGURACIÓN
# ============================================
# Instalar: pip install colorlog structlog
colorlog>=6.6.0                 # Logs con colores en consola
structlog>=23.1.0               # Logging estructurado
python-dotenv>=1.0.0            # Variables de entorno desde .env

# ============================================
# DESARROLLO Y TESTING
# ============================================
# Instalar: pip install pytest black flake8
pytest>=7.0.0                   # Framework de testing
pytest-cov>=4.0.0               # Cobertura de tests
black>=22.0.0                   # Formateador de código Python
flake8>=5.0.0                   # Linter para calidad de código
typeguard>=3.0.0                # Validación de tipos en runtime

# ============================================
# UTILIDADES ADICIONALES
# ============================================
# Instalar: pip install requests tempfile2
requests>=2.28.0                # Cliente HTTP para APIs
tempfile2>=0.1.0                # Manejo seguro de archivos temporales

# ============================================
# DEPENDENCIAS OPCIONALES (FUTURAS)
# ============================================
# Descomenta según necesidades:

# === API AVANZADA ===
# pip install fastapi uvicorn pydantic
# fastapi>=0.95.0               # API moderna más rápida que Flask
# uvicorn>=0.20.0               # Servidor ASGI para FastAPI
# pydantic>=1.10.0              # Validación de datos con tipos

# === BASE DE DATOS ===
# pip install sqlalchemy alembic
# sqlalchemy>=2.0.0             # ORM para bases de datos
# alembic>=1.9.0                # Migraciones de BD

# === PROCESAMIENTO LATEX ===
# pip install pylatex
# pylatex>=1.4.2                # Generación programática de LaTeX

# === COMPRESIÓN ===
# pip install zipfile36
# zipfile36>=0.1.3              # Manejo avanzado de archivos ZIP

# ====================================================================
# COMANDOS ÚTILES DESPUÉS DE INSTALACIÓN
# ====================================================================
#
# VERIFICAR INSTALACIÓN:
# python -c "import flask; print('Flask OK')"
# python -c "import regex; print('Regex OK')"
# python -c "import flask_login; print('Flask-Login OK')"
# 
# EJECUTAR APLICACIÓN:
# python app.py
#
# EJECUTAR TESTS:
# pytest test_*.py
#
# FORMATEAR CÓDIGO:
# black *.py
# flake8 *.py
#
# ====================================================================